llm/aliyun/qwen/requirements.txt
python-dotenv
langchain
dashscope
